+++ include/QLearning.h
@@ -22,7 +22,7 @@
 class QLearning {
  public:
   QLearning(SimInterface* sim, double learning_rate, double decay,
-            int sim_depth, double epsilon_init = 1.0,
+            int64_t sim_depth, double epsilon_init = 1.0,
             double epsilon_final = 0.1, uint64_t seed = random_device{}())
       : sim(sim),
         learning_rate(learning_rate),
@@ -37,43 +37,45 @@ class QLearning {
 
   /// @brief Train the Q-learning model on the given belief until improvement
   /// across the belief is less than epsilon or max_episodes is reached.
-  void Train(const BeliefParticles& belief, int max_episodes, int episode_size,
-             int num_sims, double epsilon, ostream& os = cout);
+  void Train(const BeliefParticles& belief, int64_t max_episodes,
+             int64_t episode_size, int64_t num_sims, double epsilon,
+             ostream& os = cout);
 
   /// @brief Return the estimated Q-value for the given state index.
-  double EstimateValue(int state, int n_sims);
+  double EstimateValue(int64_t state, int64_t n_sims);
 
   /// @brief Use the update equation for the q value associated with given state
   /// index and action index. Updates `q_table` internally.
-  void UpdateQValue(int state, int action, double reward, int next_state);
+  void UpdateQValue(int64_t state, int64_t action, double reward,
+                    int64_t next_state);
 
   /// @brief Return the current q value associated with state index and
   /// action index.
-  double GetQValue(int state, int action);
+  double GetQValue(int64_t state, int64_t action);
 
   /// @brief Return the maximum Q-value for the state index across all actions,
   /// and the best action index.
-  tuple<double, int> MaxQ(int state);
+  tuple<double, int64_t> MaxQ(int64_t state);
 
   /// @brief Choose the index of the action to take in the state index. Chooses
   /// a random action with probability epsilon, otherwise chooses the
   /// current best action.
-  int ChooseAction(int state);
+  int64_t ChooseAction(int64_t state);
 
  private:
   SimInterface* sim;
-  int n_sims;
+  int64_t n_sims;
   double learning_rate;
   double decay;
-  int sim_depth;
+  int64_t sim_depth;
   double epsilon_init;
   double epsilon_final;
   double discount;
   double epsilon;
-  unordered_map<int, vector<double>> q_table;
+  unordered_map<int64_t, vector<double>> q_table;
   mutable std::mt19937_64 rng;
 
-  unordered_map<int, vector<double>>::iterator GetQTableRow(int state);
+  unordered_map<int64_t, vector<double>>::iterator GetQTableRow(int64_t state);
 
   void DecayParameters();
 };
