+++ src/QLearning.cpp
@@ -3,8 +3,8 @@
 #include <algorithm>
 #include <limits>
 
-std::unordered_map<int, std::vector<double>>::iterator QLearning::GetQTableRow(
-    int state) {
+std::unordered_map<int64_t, std::vector<double>>::iterator
+QLearning::GetQTableRow(int64_t state) {
   const auto row = q_table.find(state);
   if (row != q_table.end()) return row;
   // Initialise row
@@ -12,12 +12,12 @@ std::unordered_map<int, std::vector<double>>::iterator QLearning::GetQTableRow(
       .first;
 }
 
-double QLearning::EstimateValue(int stateInit, int n_sims) {
-  for (int i = 0; i < n_sims; i++) {
-    int state = stateInit;
-    int depth = 0;
+double QLearning::EstimateValue(int64_t stateInit, int64_t n_sims) {
+  for (int64_t i = 0; i < n_sims; i++) {
+    int64_t state = stateInit;
+    int64_t depth = 0;
     while (depth < sim_depth) {
-      const int action = ChooseAction(state);
+      const int64_t action = ChooseAction(state);
       const auto [stateNext, o, reward, done] = sim->Step(state, action);
       ++depth;
 
@@ -40,26 +40,26 @@ void QLearning::DecayParameters() {
   learning_rate *= f;
 }
 
-void QLearning::UpdateQValue(int state, int action, double reward,
-                             int next_state) {
+void QLearning::UpdateQValue(int64_t state, int64_t action, double reward,
+                             int64_t next_state) {
   const double old_val = GetQValue(state, action);
   const double new_val = reward + discount * get<0>(MaxQ(next_state));
   const double new_Q = (1 - learning_rate) * old_val + learning_rate * new_val;
   q_table[state][action] = new_Q;
 }
 
-double QLearning::GetQValue(int state, int action) {
+double QLearning::GetQValue(int64_t state, int64_t action) {
   const auto row = GetQTableRow(state);
   return row->second.at(action);
 }
 
-tuple<double, int> QLearning::MaxQ(int state) {
+tuple<double, int64_t> QLearning::MaxQ(int64_t state) {
   const auto row = GetQTableRow(state);
   const auto best = max_element(row->second.cbegin(), row->second.cend());
   return make_tuple(*best, best - row->second.cbegin());
 }
 
-int QLearning::ChooseAction(int state) {
+int64_t QLearning::ChooseAction(int64_t state) {
   // check if we should explore randomly
   uniform_real_distribution<double> unif(0, 1);
   const double u = unif(rng);
@@ -72,16 +72,16 @@ int QLearning::ChooseAction(int state) {
   return get<1>(MaxQ(state));
 }
 
-void QLearning::Train(const BeliefParticles& belief, int max_episodes,
-                      int episode_size, int num_sims, double epsilon,
+void QLearning::Train(const BeliefParticles& belief, int64_t max_episodes,
+                      int64_t episode_size, int64_t num_sims, double epsilon,
                       ostream& os) {
   double improvement = numeric_limits<double>::infinity();
   double avg_curr = -numeric_limits<double>::infinity();
-  int i_episode = 0;
+  int64_t i_episode = 0;
   while (improvement > epsilon && i_episode < max_episodes) {
     os << "------ Episode: " << i_episode << " ------" << endl;
     double ep_value = 0.0;
-    for (int i = 0; i < episode_size; ++i) {
+    for (int64_t i = 0; i < episode_size; ++i) {
       double sum = 0.0;
       for (const auto& state : belief.GetParticles())
         sum += EstimateValue(state, num_sims);
